refactor(ricer::manager)!: move TOML handler code into `src/manager/toml.rs`
I think this will improve the overall readability of the code by separating it
out into a smaller private module.
