+++ src/manager.rs
@@ -8,12 +8,14 @@
 
 mod error;
 mod locator;
+mod toml;
 
 #[doc(inline)]
 pub use error::*;
 pub use locator::*;
+pub use toml::*;
 
-use crate::config::{CommandHook, ConfigEntry, Repository, Toml, TomlError};
+use crate::config::Toml;
 
 use log::debug;
 use mkdirp::mkdirp;
@@ -190,112 +192,5 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.doc)
-    }
-}
-
-/// TOML serialization and deserialization manager.
-///
-/// Interface to simplify serialization and deserialization of parsed TOML data.
-///
-/// # See also
-///
-/// - [`Toml`]
-pub trait TomlManager: fmt::Debug {
-    type Entry: ConfigEntry;
-
-    fn get(&self, doc: &Toml, key: &str) -> Result<Self::Entry, TomlError>;
-    fn add(&self, doc: &mut Toml, entry: Self::Entry) -> Result<Option<Self::Entry>, TomlError>;
-    fn remove(&self, doc: &mut Toml, key: &str) -> Result<Self::Entry, TomlError>;
-    fn rename(&self, doc: &mut Toml, from: &str, to: &str) -> Result<Self::Entry, TomlError>;
-    fn location<'cfg>(&self, locator: &'cfg impl Locator) -> &'cfg Path;
-}
-
-/// Repository data configuration management.
-///
-/// Handles serialization and deserialization of repository settings.
-/// Repository settings are held within the "repos" section of a
-/// configuration file.
-///
-/// # Invariants
-///
-/// Will preserve existing formatting of configuration file if any.
-///
-/// # See also
-///
-/// - [`Toml`]
-/// - [`Repository`]
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct RepositoryData;
-
-impl TomlManager for RepositoryData {
-    type Entry = Repository;
-
-    fn get(&self, doc: &Toml, key: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.get("repos", key.as_ref())?;
-        Ok(Repository::from(entry))
-    }
-
-    fn add(&self, doc: &mut Toml, entry: Self::Entry) -> Result<Option<Self::Entry>, TomlError> {
-        let entry = doc.add("repos", entry.to_toml())?.map(Repository::from);
-        Ok(entry)
-    }
-
-    fn remove(&self, doc: &mut Toml, key: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.remove("repos", key.as_ref())?;
-        Ok(Repository::from(entry))
-    }
-
-    fn rename(&self, doc: &mut Toml, from: &str, to: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.rename("repos", from.as_ref(), to.as_ref())?;
-        Ok(Repository::from(entry))
-    }
-
-    fn location<'cfg>(&self, locator: &'cfg impl Locator) -> &'cfg Path {
-        locator.repos_config()
-    }
-}
-
-/// Command hook configuration management.
-///
-/// Handles serialization and deserialization of command hook settings.
-/// Command hook settings are held within the "hooks" section of a
-/// configuration file.
-///
-/// # Invariants
-///
-/// Will preserve existing formatting of configuration file if any.
-///
-/// # See also
-///
-/// - [`Toml`]
-/// - [`CommandHook`]
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct CommandHookData;
-
-impl TomlManager for CommandHookData {
-    type Entry = CommandHook;
-
-    fn get(&self, doc: &Toml, key: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.get("hooks", key.as_ref())?;
-        Ok(CommandHook::from(entry))
-    }
-
-    fn add(&self, doc: &mut Toml, entry: Self::Entry) -> Result<Option<Self::Entry>, TomlError> {
-        let entry = doc.add("hooks", entry.to_toml())?.map(CommandHook::from);
-        Ok(entry)
-    }
-
-    fn remove(&self, doc: &mut Toml, key: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.remove("hooks", key.as_ref())?;
-        Ok(CommandHook::from(entry))
-    }
-
-    fn rename(&self, doc: &mut Toml, from: &str, to: &str) -> Result<Self::Entry, TomlError> {
-        let entry = doc.rename("hooks", from.as_ref(), to.as_ref())?;
-        Ok(CommandHook::from(entry))
-    }
-
-    fn location<'cfg>(&self, locator: &'cfg impl Locator) -> &'cfg Path {
-        locator.hooks_config()
     }
 }
